Validation plot is now prepared to plot
from every model. You have to choose the model and the folder.
Added str_of_fixed length to aux_fcn, it generates a string of fixed length from a int, padding with zeros from the left.
format_predictions, in aux_fcn is simplified, now it only needs de name of the model, file and session and it generates de corresponding .txt.